requirements.txt
requests
minimalmodbus
timeloop
datetime
pyyaml